# [CONSVC-1982] refactor: rewrite Merino middleware as pure ASGI middleware