Rewrite middlewares as pure ASGI middlewares to address the performance overheads with Starlette-based middlewares.

This is the first part for:
- geolocation
- user agent
- logging

Will convert the remainder (metrics and feature flags) in follow-ups. 

This closes [CONSVC-1982](https://mozilla-hub.atlassian.net/browse/CONSVC-1982).