@ncloudioj 2022-10-03
Load test comparision (before/after)

Test environment: running Merino-py (`MERINO_LOGGING__FORMAT=mozlog MERINO_LOGGING__LEVEL=INFO make run`) in the single-threaded mode on a 2021 MBP M1 Max.

Load tests from [merino-explorations](https://github.com/quiiver/merino-explorations). Using 100 and 200 VUs (virtual users).

## 100 VUs
Before: 775 rps & 160 ms (P95 latency)
After: 2500 rps & 52 ms

## 200 VUs
Before: 666 rps & 352 ms
After: 2537 rps & 70 ms

It's about 3-4x faster in both throughput and latency.